--- a/src/Nethermind/Nethermind.Core.Test/Caching/LruCacheTests.cs
+++ b/src/Nethermind/Nethermind.Core.Test/Caching/LruCacheTests.cs
@@ -1,4 +1,4 @@
-﻿//  Copyright (c) 2018 Demerzel Solutions Limited
+//  Copyright (c) 2018 Demerzel Solutions Limited
 //  This file is part of the Nethermind library.
 // 
 //  The Nethermind library is free software: you can redistribute it and/or modify
@@ -14,6 +14,7 @@
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with the Nethermind. If not, see <http://www.gnu.org/licenses/>.
 
+using System;
 using FluentAssertions;
 using Nethermind.Core.Caching;
 using Nethermind.Core.Test.Builders;
@@ -22,12 +23,12 @@ using NUnit.Framework;
 
 namespace Nethermind.Core.Test.Caching
 {
-    [TestFixture]
-    public class LruCacheTests
+    [TestFixture(typeof(LruCache<Address, Account>))]
+    public class LruCacheTests<TCache>
     {
-        private ICache<Address, Account> Create()
+        private static ICache<Address, Account> Create()
         {
-            return new LruCache<Address, Account>(Capacity, "test");
+            return (ICache<Address, Account>) Activator.CreateInstance(typeof(TCache), Capacity, "test");
         }
         
         private const int Capacity = 16;