@Scooletz 2020-11-19
@tkstanczak I flushed some estimations for sizes that we can discuss taking into consideration count of items and caching strategies (where it's cleared etc.)