@Scooletz 2020-11-20
@tkstanczak added modified benchmark and a benchmark result